Code Cleaning: Code Reduce & Code Readability
