+++ lib/constants/routes/routes.dart
@@ -1,23 +1,24 @@
 import "package:flutter/material.dart";
 // Admin Pages
 import 'package:rafay_portfolio/admin/frontend/pages/auth/loginPage.dart';
+import 'package:rafay_portfolio/user/frontend/pages/aboutme/aboutMe.dart';
+import 'package:rafay_portfolio/user/frontend/pages/blogs/displayblogs.dart';
+import 'package:rafay_portfolio/user/frontend/pages/error/page404.dart';
+import 'package:rafay_portfolio/user/frontend/pages/experiences/ProjectGallery.dart';
+import 'package:rafay_portfolio/user/frontend/pages/experiences/resume.dart';
+import 'package:rafay_portfolio/user/frontend/screens/contentMe_page.dart';
 
 // user Page
-import 'package:rafay_portfolio/user/frontend/screens/NotFound_page.dart';
-import 'package:rafay_portfolio/user/frontend/screens/ProjectGalley_Page.dart';
-import 'package:rafay_portfolio/user/frontend/screens/Resume_page.dart';
-import 'package:rafay_portfolio/user/frontend/screens/aboutme_page.dart';
-import 'package:rafay_portfolio/user/frontend/screens/blog_page.dart';
-import 'package:rafay_portfolio/user/frontend/screens/contentMe_page.dart';
+
 import 'package:rafay_portfolio/user/frontend/screens/home_page.dart';
 
 final Map<String, WidgetBuilder> appRoutes = {
   '/home': (context) => const HomePage(),
-  '/projects': (context) => const ProjectGalleryPage(),
-  '/resume': (context) => const ResumePage(),
+  '/projects': (context) => const ProjectGallery(),
+  '/resume': (context) => const Resume(),
   '/contact': (context) => const ContactMePage(),
-  '/blog': (context) => const BlogPage(),
-  '/aboutme': (context) => const AboutMePage(),
-  '/404': (context) => const NotFoundPage(),
+  '/blog': (context) => const DisplayBlog(),
+  '/aboutme': (context) => const AboutMe(),
+  '/404': (context) => const Page404(),
   '/admin': (context) => const LoginAdmin(),
 };
