+++ lib/user/frontend/pages/experiences/resume.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:rafay_portfolio/constants/screensSize/screentype.dart';
 import 'package:rafay_portfolio/user/frontend/class/jobExperienceData.dart';
-import 'package:rafay_portfolio/user/frontend/widgets/ExperiencesCard.dart';
+import 'package:rafay_portfolio/user/frontend/pages/experiences/widget/ExperiencesCard.dart';
 import 'package:rafay_portfolio/user/frontend/widgets/animatedtext.dart';
 import 'package:timeline_tile/timeline_tile.dart';
