+++ lib/user/frontend/pages/experiences/widget/ExperiencesCard.dart
@@ -1,8 +1,9 @@
 // // ignore_for_file: file_names
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:rafay_portfolio/constants/screensSize/screentype.dart';
 import 'package:rafay_portfolio/user/frontend/class/jobExperienceData.dart';
-import 'package:rafay_portfolio/user/frontend/views/ProjectGallery.dart';
+
 import 'package:rafay_portfolio/user/frontend/widgets/HoverChip.dart';
 import 'package:rafay_portfolio/user/frontend/widgets/textstyle.dart';
