--- lib/user/frontend/screens/NotFound_page.dart
@@ -1,26 +1,0 @@
-// ignore_for_file: file_names
-
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-
-import 'package:rafay_portfolio/user/frontend/views/page404.dart';
-
-class NotFoundPage extends StatefulWidget {
-  const NotFoundPage({super.key});
-
-  @override
-  State<NotFoundPage> createState() => _NotFoundPageState();
-}
-
-class _NotFoundPageState extends State<NotFoundPage> {
-  @override
-  Widget build(BuildContext context) {
-    SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
-      statusBarColor: Theme.of(context).colorScheme.background,
-    ));
-    return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.background,
-      body: const Page404(),
-    );
-  }
-}
