--- lib/user/frontend/screens/Resume_page.dart
@@ -1,19 +1,0 @@
-// ignore: file_names
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:rafay_portfolio/user/frontend/views/resume.dart';
-
-class ResumePage extends StatelessWidget {
-  const ResumePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
-      statusBarColor: Theme.of(context).colorScheme.background,
-    ));
-    return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.background,
-      body: const Resume(),
-    );
-  }
-}
