--- lib/user/frontend/screens/blog_page.dart
@@ -1,23 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:rafay_portfolio/user/frontend/views/displayblogs.dart';
-
-class BlogPage extends StatefulWidget {
-  const BlogPage({super.key});
-
-  @override
-  State<BlogPage> createState() => _BlogPageState();
-}
-
-class _BlogPageState extends State<BlogPage> {
-  @override
-  Widget build(BuildContext context) {
-    SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
-      statusBarColor: Theme.of(context).colorScheme.background,
-    ));
-    return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.background,
-      body: const DisplayBlog(),
-    );
-  }
-}
