--- lib/user/frontend/screens/certification_page.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CertificationPage extends StatelessWidget {
-  const CertificationPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold();
-  }
-}
