+++ lib/user/frontend/screens/contentMe_page.dart
@@ -1,16 +1,11 @@
 // ignore: file_names
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:rafay_portfolio/user/frontend/views/ContactMe.dart';
+import 'package:rafay_portfolio/user/frontend/pages/contact_me/ContactMe.dart';
 
-class ContactMePage extends StatefulWidget {
+class ContactMePage extends StatelessWidget {
   const ContactMePage({super.key});
 
-  @override
-  State<ContactMePage> createState() => _ContactMePageState();
-}
-
-class _ContactMePageState extends State<ContactMePage> {
   @override
   Widget build(BuildContext context) {
     SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
