+++ lib/user/frontend/screens/home_page.dart
@@ -3,13 +3,13 @@ import 'package:crystal_navigation_bar/crystal_navigation_bar.dart';
 import 'package:flutter/services.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:iconly/iconly.dart';
+import 'package:rafay_portfolio/user/frontend/pages/aboutme/aboutMe.dart';
+import 'package:rafay_portfolio/user/frontend/pages/blogs/displayblogs.dart';
+import 'package:rafay_portfolio/user/frontend/pages/contact_me/ContactMe.dart';
+import 'package:rafay_portfolio/user/frontend/pages/experiences/ProjectGallery.dart';
+import 'package:rafay_portfolio/user/frontend/pages/experiences/resume.dart';
 
-import 'package:rafay_portfolio/user/frontend/views/landing_page.dart';
-import 'package:rafay_portfolio/user/frontend/screens/Resume_page.dart';
-import 'package:rafay_portfolio/user/frontend/screens/aboutme_page.dart';
-import 'package:rafay_portfolio/user/frontend/screens/blog_page.dart';
-import 'package:rafay_portfolio/user/frontend/screens/contentMe_page.dart';
-import 'package:rafay_portfolio/user/frontend/screens/ProjectGalley_Page.dart';
+import 'package:rafay_portfolio/user/frontend/pages/landing/landing_page.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({Key? key}) : super(key: key);
@@ -23,11 +23,11 @@ class _HomePageState extends State<HomePage> {
 
   final List<Widget> _pages = [
     const LandingPage(),
-    const ResumePage(),
-    const ProjectGalleryPage(),
-    const AboutMePage(),
-    const BlogPage(),
-    const ContactMePage(),
+    const Resume(),
+    const ProjectGallery(),
+    const AboutMe(),
+    const DisplayBlog(),
+    const ContactMe()
   ];
 
   @override
