+++ pubspec.yaml
@@ -40,6 +40,7 @@ dependencies:
   connectivity_plus: ^5.0.2
   intl: ^0.19.0
   scroll_to_index: ^3.0.1
+  cached_network_image: ^3.3.1
 dev_dependencies:
   flutter_test:
     sdk: flutter
